+++ R/AIC.list.R
@@ -27,7 +27,7 @@ AIC.list <- function(object, ..., assess.best=TRUE) {
           if ("indentation" %in% names(ret)) {
             ret$indentation <- ret$indentation + 1
           } else {
-            stop("Unknown way to get a data.frame without indentation set.  This is likely a bug.")
+            stop("Unknown way to get a data.frame without indentation set.  This is likely a bug.") # nocov
           }
         }
       }